+++ srcs/screen_management.c
@@ -1,12 +1,12 @@
 #include "../includes/fdf.h"
 #include "../libft/include/libft.h"
 
-void	draw_line(float *from, float *to, int **screen)
+void	draw_line(double *from, double *to, int **screen)
 {
-	float dx;
-	float dy;
-	float D;
-	float y;
+	double dx;
+	double dy;
+	double D;
+	double y;
 	int x;
 
 	dx = from[0] - to[0];
@@ -27,7 +27,7 @@ void	draw_line(float *from, float *to, int **screen)
 
 }
 
-void put_to_screen_string(float *tab, int **screen)
+void put_to_screen_string(double *tab, int **screen)
 {
 	int where_to;
